requirements.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.2.0
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
h11==0.14.0
httptools==0.6.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
sympy==1.13.3
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.13
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
